+++ README.md
@@ -2,6 +2,16 @@
 
 An example workflow for automatically creating a video transcript with show notes using ChatGPT and Whisper.
 
+## Project Structure
+
+- `autogen.js` - Main entry point for the CLI
+- `utils/index.js` - Utility functions to get the model and run common file operations
+- `commands` - Directory for commands
+  - `commands/processVideo.js` - Handles processing of a single video
+  - `commands/processPlaylist.js` - Handles processing of a playlist
+  - `commands/processUrlsFile.js` - Handles processing of a file with URLs
+  - `commands/processRssFeed.js` - Handles processing of an RSS feed
+
 ## Setup
 
 ### Install Local Dependencies
@@ -27,37 +37,25 @@ make -C whisper.cpp
 
 ## Run Autogen Bash Scripts
 
-Run on a single YouTube video.
+
 
 ```bash
-# short one minute video
+# Run on a single YouTube video (short one minute video)
 ./autogen.sh --video "https://www.youtube.com/watch?v=jKB0EltG9Jo"
 
-# longer 30 minute video
+# Run on a single YouTube video (longer 30 minute video)
 ./autogen.sh --video "https://www.youtube.com/watch?v=QhXc9rVLVUo"
-```
 
-Run on a single audio file.
-
-```bash
+# Run on a single audio file
 ./autogen.sh --audio "https://media.transistor.fm/d1d18d2d/449ace19.mp3"
-```
 
-Run on multiple YouTube videos in a playlist.
-
-```bash
+# Run on multiple YouTube videos in a playlist
 ./autogen.sh --playlist "https://www.youtube.com/playlist?list=PLCVnrVv4KhXMh4DQBigyvHSRTf2CSj129"
-```
 
-Run on an arbitrary list of URLs in `urls.md`.
-
-```bash
+# Run on an arbitrary list of URLs in `urls.md`
 ./autogen.sh --urls urls.md
-```
 
-Run on a local video file.
-
-```bash
+# Run on a local video file
 ./autogen.sh --file content/video.mkv
 ```
